Forgot to remove a simple debug

+++ src/main/java/com/wynntils/core/framework/rendering/ScreenRenderer.java
@@ -545,10 +545,10 @@ public class ScreenRenderer {
               yMax  = Math.max(y1, y2) + drawingOrigin.y;
 
         GlStateManager.glBegin(GL_QUADS);
-        GlStateManager.glVertex3f((int)xMin, (int)yMin, 0);
-        GlStateManager.glVertex3f((int)xMin, (int)yMax, 0);
-        GlStateManager.glVertex3f((int)xMax, (int)yMax, 0);
-        GlStateManager.glVertex3f((int)xMax, (int)yMin, 0);
+        GlStateManager.glVertex3f(xMin, yMin, 0);
+        GlStateManager.glVertex3f(xMin, yMax, 0);
+        GlStateManager.glVertex3f(xMax, yMax, 0);
+        GlStateManager.glVertex3f(xMax, yMin, 0);
         GlStateManager.glEnd();
         GlStateManager.enableTexture2D();
         GlStateManager.color(1f,1f,1f,1f);
